+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - Fix(tui): fix netease lyric download error.
 - Fix(tui): fix migu search/lyric download error.
 - Fix(tui): in tag editor, dont save the track on lyric delete (now only on "Save" action is the change actually persistet).
+- Fix(tui): on tag editor close, force reload of lyrics if the current track was modified.
 - Fix(server): on rusty backend with `soundtouch`, fix a issue causing seemingly long delay between tempo changes.
 - Fix(server): on rusty backend with `soundtouch`, fix not being able to build on "eager" linkers (ex windows-msvc and windows-gnu)
 - Feat: add ability to control startup playing state behavior.